+++ packages/cli/src/commands/__tests__/lint.test.ts
@@ -81,6 +81,7 @@ describe('lint', () => {
         output: { stylish: '<stdout>' },
         ignoreUnknownFormat: false,
         failOnUnmatchedGlobs: false,
+        showDocumentationUrl: false,
       });
     });
   });
@@ -94,6 +95,7 @@ describe('lint', () => {
       output: { stylish: '<stdout>' },
       ignoreUnknownFormat: false,
       failOnUnmatchedGlobs: false,
+      showDocumentationUrl: false,
     });
   });
 
@@ -106,6 +108,7 @@ describe('lint', () => {
       output: { stylish: '<stdout>' },
       ignoreUnknownFormat: false,
       failOnUnmatchedGlobs: false,
+      showDocumentationUrl: false,
     });
   });
 
@@ -118,6 +121,7 @@ describe('lint', () => {
       output: { json: '<stdout>' },
       ignoreUnknownFormat: false,
       failOnUnmatchedGlobs: false,
+      showDocumentationUrl: false,
     });
   });
 
@@ -184,6 +188,7 @@ describe('lint', () => {
       output: { stylish: '<stdout>' },
       ignoreUnknownFormat: true,
       failOnUnmatchedGlobs: false,
+      showDocumentationUrl: false,
     });
   });
 
@@ -195,6 +200,7 @@ describe('lint', () => {
       output: { stylish: '<stdout>' },
       ignoreUnknownFormat: false,
       failOnUnmatchedGlobs: true,
+      showDocumentationUrl: false,
     });
   });
 
@@ -244,13 +250,13 @@ describe('lint', () => {
     expect(process.stderr.write).nthCalledWith(2, `Error #1: ${chalk.red('some unhandled exception')}\n`);
     expect(process.stderr.write).nthCalledWith(
       3,
-      expect.stringContaining(`packages/cli/src/commands/__tests__/lint.test.ts:236`),
+      expect.stringContaining(`packages/cli/src/commands/__tests__/lint.test.ts:242`),
     );
 
     expect(process.stderr.write).nthCalledWith(4, `Error #2: ${chalk.red('another one')}\n`);
     expect(process.stderr.write).nthCalledWith(
       5,
-      expect.stringContaining(`packages/cli/src/commands/__tests__/lint.test.ts:237`),
+      expect.stringContaining(`packages/cli/src/commands/__tests__/lint.test.ts:243`),
     );
 
     expect(process.stderr.write).nthCalledWith(6, `Error #3: ${chalk.red('original exception')}\n`);