+++ packages/functions/src/xor.ts
@@ -1,10 +1,8 @@
 import { createRulesetFunction, IFunctionResult } from '@stoplight/spectral-core';
-import { printValue } from '@stoplight/spectral-runtime';
-
 import { optionSchemas } from './optionSchemas';
 
 export type Options = {
-  /** test to verify if one (but not all) of the provided keys are present in object */
+  /** test to verify if exactly one of the provided keys are present in object */
   properties: string[];
 };
 
@@ -16,20 +14,32 @@ export default createRulesetFunction<Record<string, unknown>, Options>(
     options: optionSchemas.xor,
   },
   function xor(targetVal, { properties }) {
+    if (properties.length < 2) return;
+    // At least two but no maximum limit on number of properties
+
     const results: IFunctionResult[] = [];
 
-    const intersection = Object.keys(targetVal).filter(key => properties.includes(key));
+    const intersection = Object.keys(targetVal).filter(value => -1 !== properties.indexOf(value));
+    if (intersection.length == 0) {
+      if (properties.length > 4) {
+        // List first three properties and remaining count in error message
+        const shortprops = properties.slice(0, 3);
+        const count = String(properties.length - 3) + ' other properties must be defined';
+        results.push({
+          message: 'At least one of "' + shortprops.join('" or "') + '" or ' + count,
+        });
+      } else {
+        // List all of two to four properties directly in error message
+        results.push({
+          message: 'At least one of "' + properties.join('" or "') + '" must be defined',
+        });
+      }
+    }
 
-    if (intersection.length !== 1) {
-      const formattedProperties = properties.map(prop => printValue(prop));
-
-      const lastProperty = formattedProperties.pop();
-      let message = formattedProperties.join(', ') + (lastProperty != undefined ? ` and ${lastProperty}` : '');
-
-      message += ' must not be both defined or both undefined';
-
+    if (intersection.length > 1) {
+      // List all defined properties in error message
       results.push({
-        message,
+        message: 'Just one of "' + intersection.join('" and "') + '" must be defined',
       });
     }